# In polars_io::csv::parser::parse_lines, change unsafe block into iter_mut